Split reporter logic
